requirements.txt
ollama
langchain-core
langchain-ollama
python-dotenv
pandas
scikit-learn
langchain-google-genai
langchain-openai